addhardware: Drop 'storage format' field
Originally this made sense, as it was the only way to specify a non-default
storage format when creating new storage.

Nowadays the storage browser is a full featured storage manager... and
this field is a bit confusing WRT whether it's used for creating new
storage, or informing libvirt about an existing image's format.

Drop it from the addhardware wizard, and simplify what we show in the
details wizard as well.