details: Remove UI for vcpu pinning
This UI exposed the old NUMA allocation policy detailed in the last commit.
It's very much sub-optimal, and should be removed.

Manual cpuset configuration is also quite uncommon and not really something
worth exposing in the UI. It can easily be done from the command line
with virt-xml.

If people complain, I'd consider adding a checkbox for vcpu placement=auto,
or an option to do that by default for new VMs.