virtconvtest: Fix vmx2libvirt test
The test consists of translating VMX configuration into domain
XML and converting disks. Cool. But the disk is zipped in a file
and the test tries to string match unzipping command. Problem is,
the absolute path is passed to the unzip command which makes it
impossible for the test to succeed on other hosts.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>

(crobinso: fix whitespace)