tests: Remove storage UUID hackery
In several cases we were trying to define two different pools with
the same UUID. Libvirt used to accept that but as of a recent
commit it started rejecting it for the test driver.

The UUID stuff isn't adding anything so drop it from all the test
cases and data.