choosecd: Emit cdrom-chosen signal after the dialog has been closed
This fixes the testChooseCDHotplug UI test: the cdrom-chosen signal
opens an alert and it does not get focus because it's being opened
before the dialog is closed. As a result, the tests timeout waiting for
the alert to get focus. Emitting the signal after the closure of the
dialog fixes this issue.

Signed-off-by: Povilas Kanapickas <povilas@radix.lt>