installertreemedia: Add unattended_data to the class
unattended_data will be used when we add support for unattended
installations. It'll be passed down to the helper that's responsible for
actually generating the script and the kernel command line to be used
with it during prepare(); nevertheless, the addition does nothing right
now.

Together with the addition of the unattended_data, a setter has been
added.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>