installertreemedia: Perform a unattended installation
In case the InstallerTreeMedia has any unattended_data set, let's just
generate the install script and have everything set up for performing a
unattended installation.

path is appended to the tmpfiles as it's the generated unattended script
and it has to be cleaned up when no longer needed.

cmdline is preppended to the args gotten from _prepare_kernel_url() in
order to be used during the first OS boot (and only there!).

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>