installertreemdia: Set user specified --location as the installer URL
The location has to be passed down to the generate_install_script()
method whenever a tree based installation is being performed.

Although Libosinfo provides information about the URL to be used when
performing a tree based installation, cases where the users prefer to
use, for any specific reason, their own URLs should be covered as well.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>