unattended: Properly set the installation source
The installation source, which can be either "media" or "network",
should be properly set depending on whether we're installing from a
media or performing a tree based installation.

We have to mind here that any media that is set as "netinstall" variant
has to have its installation source set accordingly.

An different approach for this would be to generate the script
specifically from the media whenever a media is passed. However, IMO, it
would just add two new codepaths and having the media passed around
without a real gain against the method chosen.

Signed-off-by: Fabiano Fidêncio <fidencio@redhat.com>